+++ applications/vex2difx/trunk/vexdatamodel/vex_setup.h
@@ -52,7 +52,7 @@ public:
 	void setPhaseCalInterval(int phaseCalIntervalMHz);
 	void selectTones(enum ToneSelection selection, double guardBandMHz);
 	bool usesFormat(enum VexStream::DataFormat format) const;
-	int nStream() const { return streams.size(); }
+	size_t nStream() const { return streams.size(); }
 	int nRecordChan() const;
 	unsigned int getBits() const;
 	unsigned int getMinBits() const;
